Extract class to test proposal reactivate.
